upgrades: remove migration that waits for schema changes
We can also remove some skipped tests, since they no longer apply.

Release note: None
