+++ pkg/upgrade/upgrades/BUILD.bazel
@@ -36,7 +36,6 @@ go_library(
         "tenant_table_migration.go",
         "upgrades.go",
         "wait_for_del_range_in_gc_job.go",
-        "wait_for_schema_changes.go",
         "web_sessions_table_user_id_migration.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/upgrade/upgrades",
@@ -123,7 +122,6 @@ go_test(
         "system_statistics_activity_test.go",
         "tenant_table_migration_test.go",
         "wait_for_del_range_in_gc_job_test.go",
-        "wait_for_schema_changes_test.go",
         "web_sessions_table_user_id_migration_test.go",
     ],
     args = ["-test.timeout=895s"],
@@ -185,7 +183,6 @@ go_test(
         "//pkg/util/intsets",
         "//pkg/util/leaktest",
         "//pkg/util/log",
-        "//pkg/util/protoutil",
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "@com_github_cockroachdb_errors//:errors",
